requirements.txt
pika
cassandra-driver
psutil
pandas
streamlit